requirements.txt
backcall==0.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
decorator==4.4.0
idna==2.8
imutils==0.5.3
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
MouseInfo==0.0.4
mss==4.0.3
numpy==1.17.2
opencv-python==4.1.1.26
parso==0.4.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
PyAutoGUI==0.9.47
PyGetWindow==0.0.7
Pygments==2.4.2
PyMsgBox==1.0.7
pyperclip==1.7.0
pyreadline==2.1
PyRect==0.1.4
PyScreeze==0.1.22
PyTweening==1.0.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
